silence warning of unsigned -> int

+++ tests/write.cpp
@@ -221,7 +221,7 @@ int main(){
     strings.push_back(string);
 
     std::array<int, 4> arrayTest = {0, 0, 2, 3};
-    std::array<int, 4> arrayTest2 = {4, 4, 2*i};
+    std::array<int, 4> arrayTest2 = {4, 4, 2*static_cast<int>(i)};
     NotSoSimpleStruct a;
     a.data.p = arrayTest2;
     auto array = ExampleWithArray(a, arrayTest, arrayTest, arrayTest, arrayTest);
